server/requirements.txt
facenet-pytorch==2.6.0
fastapi==0.112.2
fiftyone==0.25.0
fire==0.6.0
Flask==3.0.3
gunicorn==22.0.0
httptools==0.6.1
Keras-Applications==1.0.8
keras-facenet==0.3.2
keras-vggface==0.6
natsort==8.4.0
pipdeptree==2.21.0
pycocotools==2.0.7
PySocks==1.7.1
python-dotenv==1.0.1
retina-face==0.0.17
tf_keras==2.16.0
ultralytics==8.1.29
#uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
#motor~=3.4.0
motor~=3.3.1 ##################
pymongo~=4.5.0 ################
numpy~=1.26.4
uvicorn~=0.30.3
opencv-python~=4.9.0.80
requests~=2.32.3
torch~=2.2.1
torchvision==0.17.2
pillow~=10.2.0
scikit-learn~=1.5.0
pydantic~=2.8.2
ffmpeg-python~=0.2.0
#git+https://github.com/ClementPinard/FlowNetPytorch.git
transformers==4.41.2 # Added for CLIP